requirements.txt
astroid==3.3.9
azure-core==1.33.0
azure-functions==1.21.3
azure-storage-blob==12.25.1
azure-storage-queue==12.12.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
coverage==7.8.0
cryptography==44.0.2
dill==0.3.9
flake8==7.1.2
idna==3.10
iniconfig==2.1.0
isodate==0.7.2
isort==6.0.1
mccabe==0.7.0
mypy==1.15.0
mypy-extensions==1.0.0
packaging==24.2
platformdirs==4.3.7
pluggy==1.5.0
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
pylint==3.3.6
pytest==8.3.5
pytest-cov==6.1.1
requests==2.32.3
six==1.17.0
tomlkit==0.13.2
typing_extensions==4.13.0
urllib3==2.3.0
